moved cycle checker to config file generation
