+++ src/main/scala/rise/autotune/configFileGeneration.scala
@@ -15,6 +15,8 @@ object configFileGeneration {
 
     val parametersWDC = distributeConstraints(p, c)
 
+    assert(check_no_cycle(parametersWDC))
+
     // number of samples for design of experiment phase
     val doe = p.size * 10
 
@@ -296,6 +298,33 @@ object configFileGeneration {
     })
 
     output.toSeq.sortBy(_.name)
+  }
+
+  // function to check cycle in parameter dependencies
+  def check_no_cycle(distribution: Map[NatIdentifier,
+    (Set[constraints.Constraint], Set[NatIdentifier])]
+                    ): Boolean = {
+
+    def check_no_cycle_rec(param: NatIdentifier,
+                           dependencies: Set[NatIdentifier],
+                           distribution: Map[NatIdentifier,
+                             (Set[constraints.Constraint], Set[NatIdentifier])]
+                          ): Boolean = {
+
+      dependencies.size match {
+        case 0 => true
+        case _ => {
+          dependencies.exists(dependency => dependency.name.equals(param.name)) match {
+            case true => false
+            case false => dependencies.forall(dependency => {
+              check_no_cycle_rec(param, distribution.apply(dependency)._2, distribution)
+            })
+          }
+        }
+      }
+    }
+
+    distribution.forall(param => check_no_cycle_rec(param._1, param._2._2, distribution))
   }
 
 }
